requirements.txt
APScheduler==3.6.3
asgiref==3.2.3
cycler==0.10.0
Cython==0.29.14
Django==3.0.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.0
object-detection-0.1==0.1
Pillow==6.2.1
PyJWT==1.7.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
tzlocal==2.0.0
